Optionally configure log template format
The `metron_agent` syslog template[1] and the `syslog-release` template[2]
are not equivalent. Also, the syslog_release one does not include the
index number of the VM.

Beause that, if one replaces the  metron_agent  syslog forwarding feature with
the syslog-release/syslog-forwarder the log message format will change.
This can cause additional parsing fail, like for instance
the filters from logsearch-for-cloudfoundry[3]

In this commit we add the optional property `syslog.log_template` to chose
different template formats, and we provide the 'default' format from
syslog-release/syslog-forwarder and the format of 'metron_agent'.

Passing a log format will allow later add additional formats like rfc5434,
making it easier to transition between then. This will enable easily
implement the solution for [4]

These are issues in all the related repos to start using a consistent
format:

 * https://github.com/cloudfoundry/syslog-release/issues/3
 * https://github.com/cloudfoundry/loggregator/issues/134
 * https://github.com/cloudfoundry-community/logsearch-for-cloudfoundry/issues/204

[1] https://github.com/cloudfoundry/loggregator/blob/develop/jobs/metron_agent/templates/syslog_forwarder.conf.erb#L53
[2] https://github.com/cloudfoundry/syslog-release/blob/master/jobs/syslog_forwarder/templates/rsyslog.conf.erb#L56
[3] https://github.com/cloudfoundry-community/logsearch-for-cloudfoundry/blob/develop/src/logsearch-config/src/logstash-filters/snippets/platform.conf#L11-L16
[4] https://github.com/cloudfoundry/syslog-release/issues/3