Add post custom config before the file exclusion
We must configure the post custom configuration before the file
exclusion rules, otherwise all the logs from blackbox, which
are marked as procid=rs2, will be dropped before reaching this config.

This makes sense with the property, as we want this config to
be the last one before the original default config of rsyslog,
and the 40-syslog-release-file-exclusion.conf does clean the logs for
the default rsyslog config.